Increased tag type to 32 bit; bumped binary grammar revision accordingly.
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@4082 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/Tag.h
@@ -58,7 +58,7 @@ namespace CG3 {
 
 	class Tag {
 	public:
-		uint16_t type;
+		uint32_t type;
 		mutable URegularExpression *regexp;
 
 		bool in_grammar;
